Fix bug whereby an attempt is made to delete nonexistent DOICollection

+++ dspace-xmlui/src/main/java/org/dspace/app/xmlui/aspect/administrative/FlowContainerUtils.java
@@ -135,7 +135,10 @@ public class FlowContainerUtils
 		}
 		else {
 			DOICollection doiCollectionObj = doiCollectionService.findByCollectionUUID(context, collectionID);
-			doiCollectionService.delete(context, doiCollectionObj);
+
+			if(doiCollectionObj != null) {
+				doiCollectionService.delete(context, doiCollectionObj);
+			}
 		}
 
 		// If they don't have a name then make it untitled.
